+++ MinecraftProject/Util/Vertex.h
@@ -14,11 +14,11 @@ struct Vertex
 	vec3 Position;
 	vec2 TexCoords;
 	vec4 Color;
-	float IndexTexture;
+	int IndexTexture;
 	vec3 Normal;
-	float Brightness;
-	Vertex(float x, float y, float z, float u, float v, float r, float g, float b, float a, uint16_t indexTexture, float normalX, float normalY, float normalZ, float brightness);
-	Vertex(const vec3& pos, float u, float v, float r, float g, float b, float a, uint16_t indexTexture, float normalX, float normalY, float normalZ, float brightness);
+	int Brightness;
+	Vertex(float x, float y, float z, float u, float v, float r, float g, float b, float a, uint16_t indexTexture, float normalX, float normalY, float normalZ, int brightness);
+	Vertex(const vec3& pos, float u, float v, float r, float g, float b, float a, uint16_t indexTexture, float normalX, float normalY, float normalZ, int brightness);
 	[[nodiscard]] vec3 GetPosition() const;
 	[[nodiscard]] vec2 GetTexCoords() const;
 	[[nodiscard]] vec4 GetColor() const;
