+++ MinecraftProject/World/Blocks/CobblestoneBlock.h
@@ -4,6 +4,6 @@
 class CobblestoneBlock final : public Block
 {
 public:
-    void GenerateTessellationData(TessellationHelper& tessellationHelper, const BlockFaces& face, const float& x, const float& y, const float& z, const float& brightness) const override;
+    void GenerateTessellationData(TessellationHelper& tessellationHelper, const BlockFaces& face, const float& x, const float& y, const float& z, const int& brightness) const override;
     CobblestoneBlock();
 };
