+++ MinecraftProject/World/Blocks/GrassBlock.h
@@ -8,6 +8,6 @@
 class GrassBlock final : public Block
 {
 public:
-    void GenerateTessellationData(TessellationHelper& tessellationHelper, const BlockFaces& face, const float& x, const float& y, const float& z, const float& brightness) const override;
+    void GenerateTessellationData(TessellationHelper& tessellationHelper, const BlockFaces& face, const float& x, const float& y, const float& z, const int& brightness) const override;
     GrassBlock();
 };
